# server/docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/reminders/preview": {
      "post": {
        "operationId": "RemindersController_preview",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PreviewReminderDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Structured preview with recurrence and first occurrence",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReminderPreviewResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid, unsupported, or past schedule"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Resolve and validate a reminder schedule without persisting it",
        "tags": [
          "Reminders"
        ]
      }
    },
    "/reminders": {
      "post": {
        "operationId": "RemindersController_create",
        "parameters": [
          {
            "name": "Prefer",
            "in": "header",
            "description": "Use return=representation to include materialized occurrences in the response",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "description": "Stable 8–128 character key reused only for the same create intent",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateReminderDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created reminder with its schedule and materialized occurrences",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreatedReminderResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Missing confirmation, invalid key, or invalid schedule"
          },
          "409": {
            "description": "Idempotency key reused with a different request"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create one explicitly confirmed reminder atomically",
        "tags": [
          "Reminders"
        ]
      }
    },
    "/parse-reminder": {
      "post": {
        "operationId": "RemindersController_parse",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ParseReminderDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Schema-validated proposal, confidence, ambiguities, warnings, and alternatives",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ParseReminderResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Parse a bounded natural-language draft into an editable proposal without creating it",
        "tags": [
          "Reminders"
        ]
      }
    },
    "/reminder-occurrences": {
      "get": {
        "operationId": "RemindersController_listOccurrences",
        "parameters": [
          {
            "name": "view",
            "required": false,
            "in": "query",
            "schema": {
              "default": "UPCOMING",
              "type": "string",
              "enum": [
                "UPCOMING",
                "OVERDUE",
                "COMPLETED",
                "ALL"
              ]
            }
          },
          {
            "name": "from",
            "required": false,
            "in": "query",
            "schema": {
              "format": "date-time",
              "type": "string"
            }
          },
          {
            "name": "to",
            "required": false,
            "in": "query",
            "schema": {
              "format": "date-time",
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 50,
              "default": 30,
              "allOf": [
                {
                  "$ref": "#/components/schemas/Object"
                }
              ]
            }
          },
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "description": "Opaque cursor returned by the previous page.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Stable time-ordered occurrence page"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List the signed-in user’s scheduled reminder occurrences",
        "tags": [
          "Reminders"
        ]
      }
    },
    "/reminders/{reminderId}": {
      "get": {
        "operationId": "RemindersController_getReminder",
        "parameters": [
          {
            "name": "reminderId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Reminder detail"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get one owned reminder with its current schedule and occurrences",
        "tags": [
          "Reminders"
        ]
      },
      "patch": {
        "operationId": "RemindersController_updateContent",
        "parameters": [
          {
            "name": "reminderId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateReminderContentDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated reminder revision"
          },
          "409": {
            "description": "Stale reminder revision or idempotency conflict"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update reminder content with optimistic concurrency",
        "tags": [
          "Reminders"
        ]
      },
      "delete": {
        "operationId": "RemindersController_deleteReminder",
        "parameters": [
          {
            "name": "reminderId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeleteReminderDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Reminder cancelled and local notifications invalidated"
          },
          "409": {
            "description": "Stale reminder revision or idempotency conflict"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Cancel and schedule purge of an owned reminder",
        "tags": [
          "Reminders"
        ]
      }
    },
    "/reminders/{reminderId}/checklist": {
      "patch": {
        "operationId": "RemindersController_replaceChecklist",
        "parameters": [
          {
            "name": "reminderId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReplaceChecklistDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated checklist template and reminder revision"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Replace the reminder checklist template and future occurrence snapshots",
        "tags": [
          "Reminders"
        ]
      }
    },
    "/reminder-occurrences/{occurrenceId}/checklist-items/{itemId}": {
      "patch": {
        "operationId": "RemindersController_toggleChecklistItem",
        "parameters": [
          {
            "name": "occurrenceId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "itemId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ToggleChecklistItemDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated occurrence checklist item"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Check or uncheck one occurrence checklist item",
        "tags": [
          "Reminders"
        ]
      }
    },
    "/reminders/{reminderId}/workflows": {
      "post": {
        "operationId": "RemindersController_createWorkflow",
        "parameters": [
          {
            "name": "reminderId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateWorkflowDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Workflow and ordered reminder steps"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a bounded sequential workflow that starts after completion",
        "tags": [
          "Reminders"
        ]
      },
      "get": {
        "operationId": "RemindersController_listWorkflows",
        "parameters": [
          {
            "name": "reminderId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List workflows bound to a reminder",
        "tags": [
          "Reminders"
        ]
      }
    },
    "/workflows/{workflowId}/lifecycle": {
      "patch": {
        "operationId": "RemindersController_updateWorkflowLifecycle",
        "parameters": [
          {
            "name": "workflowId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateWorkflowLifecycleDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Pause or resume a completion workflow with optimistic concurrency",
        "tags": [
          "Reminders"
        ]
      }
    },
    "/reminders/{reminderId}/context-triggers": {
      "post": {
        "operationId": "RemindersController_createContextTrigger",
        "parameters": [
          {
            "name": "reminderId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateContextTriggerDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Capability-aware context trigger"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create an encrypted location or keyed Wi-Fi trigger",
        "tags": [
          "Reminders"
        ]
      },
      "get": {
        "operationId": "RemindersController_listContextTriggers",
        "parameters": [
          {
            "name": "reminderId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List context triggers attached to one reminder template",
        "tags": [
          "Reminders"
        ]
      }
    },
    "/context-triggers": {
      "get": {
        "operationId": "RemindersController_listAllContextTriggers",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List the current user context triggers for device reconciliation",
        "tags": [
          "Reminders"
        ]
      }
    },
...
  },
  "info": {
    "title": "Smart Reminder API",
    "description": "Versioned API for the Smart Reminder mobile application",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "PreviewReminderDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "minLength": 1,
            "maxLength": 120
          },
          "contextNote": {
            "type": "string",
            "maxLength": 2000
          },
          "schedule": {
            "$ref": "#/components/schemas/ReminderScheduleDto"
          },
          "checklist": {
            "maxItems": 50,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CreateReminderChecklistItemDto"
            }
          }
        },
        "required": [
          "title",
          "schedule"
        ]
      },
...
      "ReminderPreviewResponseDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "contextNote": {
            "type": "object",
            "nullable": true
          },
          "schedule": {
            "$ref": "#/components/schemas/ResolvedScheduleDto"
          },
          "requiresConfirmation": {
            "type": "boolean",
            "example": true
          },
          "persisted": {
            "type": "boolean",
            "example": false
          }
        },
        "required": [
          "title",
          "schedule",
          "requiresConfirmation",
          "persisted"
        ]
      },
...
      "Object": {
        "type": "object",
        "properties": {}
      },
...
      "DeleteReminderDto": {
        "type": "object",
        "properties": {
          "expectedRevision": {
            "type": "number",
            "minimum": 1
          }
        },
        "required": [
          "expectedRevision"
        ]
      },
...
    }
  }
}